==> 0_NLP_additional.ipynb <==
# Creating extra NLP features

- **Spell error ratio** using 'autocorrect' package
- **Polarity** and **subjectivity** from TextBlob
- Train Wikipedia comments:  **['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate']**
- Readability: **Flesch-Kincaid** and **Automated Readability Index**
- **Spam** classfication probability is added using a Naive Bayes model trained w/1600-spam-corpus


# Read files

### Previous data set w/ Stanford NLP features will be loaded.

import os
import numpy as np
import pandas as pd
import time
import sys
from sklearn.linear_model import LogisticRegression
import nltk
from nltk.tokenize import RegexpTokenizer
from nltk.stem import SnowballStemmer
nltk.download('stopwords')
from nltk.corpus import stopwords
from autocorrect import spell
from textblob import TextBlob
import re, string
from readability_score.calculators.fleschkincaid import *
from readability_score.calculators.ari import *
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import accuracy_score
from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import classification_report
import re, string
try:
    sys.getwindowsversion()
except AttributeError:
    isWindows = False
else:
    isWindows = True
if isWindows:
    import win32api,win32process,win32con
    pid = win32api.GetCurrentProcessId()
    handle = win32api.OpenProcess(win32con.PROCESS_ALL_ACCESS, True, pid)
    win32process.SetPriorityClass(handle, win32process.HIGH_PRIORITY_CLASS)
regular_reviews = pd.read_csv('data/reg_reviews_NLP.csv')
not_recommended_reviews= pd.read_csv('data/not_reviews_NLP.csv')

len(regular_reviews)

len(not_recommended_reviews)

not_recommended_reviews.describe()

# spellErrorRatio function

# Create a filter for punctuations
punctuation = re.compile(r'[0-9]')

# Create a regular expression tokenizer
tokenizer = RegexpTokenizer(r'\w+')

# Get the list of stop words 
stop_words = stopwords.words('english')

def spellErrorRatio(input_text):

    # Remove the numbers
    input_text = punctuation.sub("",input_text)
    
    # Tokenize the input string
    tokens = tokenizer.tokenize(input_text)

    # Remove the stop words 
    tokens = [x for x in tokens if not x in stop_words]
    
    # Check spelling error ratio
    if len(tokens)>1:
        cntWrong = 0
        for i, item in enumerate(tokens):
            if(tokens[i]!=spell(item)):
                cntWrong += 1
    #             print(tokens[i], spell(item))  # check which word is wrong.
        return (cntWrong/len(tokens))
    else:
        return (np.nan)

# calculate spell error ratio on both files and save

t0 = time.time()
regular_reviews['spellErrorRatio'] = regular_reviews['review'].apply(lambda x: spellErrorRatio(x))
regular_reviews.to_csv('data/regular_reviews_spell.csv', index=False)
print('hours processed for one file:', (time.time()-t0)/3600)

not_recommended_reviews['spellErrorRatio'] = not_recommended_reviews['review'].apply(lambda x: spellErrorRatio(x))
not_recommended_reviews.to_csv('data/not_reviews_spell.csv', index=False)
print('total hours processed:', (time.time()-t0)/3600)

## 2. Sentiment Features using 'Text Blob'

# # for regular review file
# regular_reviews['TB_polarity'] = regular_reviews.review.apply(lambda x:TextBlob(x).polarity)
# regular_reviews['TB_subjectivity'] = regular_reviews.review.apply(lambda x:TextBlob(x).subjectivity)
# regular_reviews.to_csv('data/regular_reviews_TB.csv', index=False)

# # for not-recommended review file
# not_recommended_reviews['TB_polarity'] = not_recommended_reviews.review.apply(lambda x:TextBlob(x).polarity)
# not_recommended_reviews['TB_subjectivity'] = not_recommended_reviews.review.apply(lambda x:TextBlob(x).subjectivity)
# not_recommended_reviews.to_csv('data/not_reviews_TB.csv', index=False)

# 3. ['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate']

### Source: Kaggle/ 'Toxic Comment Classification Challenge'
Kaggle objective: Detecting toxic comments that are rude, disrespectful or otherwise likely to make someone leave a discussion.

Kaggle supporter: Conversation AI team by Jigsaw and Google.

### Train data: Wikipedia comments that are labeled by human raters for toxic behavior.  



## Training

train = pd.read_csv('data/train_Toxic.csv')
label_cols = ['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate']
train['none'] = 1-train[label_cols].max(axis=1)
COMMENT = 'comment_text'
train[COMMENT].fillna("unknown", inplace=True)
train.describe()

train.head()

## Building the model

re_tok = re.compile(f'([{string.punctuation}“”¨«»®´·º½¾¿¡§£₤‘’])')
def tokenize(s): return re_tok.sub(r' \1 ', s).split()

def pr(x,y_i, y):
    p = x[y==y_i].sum(0)
    return (p+1) / ((y==y_i).sum()+1)

def get_mdl(x,y):
    y = y.values
    r = np.log(pr(x,1,y) / pr(x,0,y))
    m = LogisticRegression(C=4, dual=True)
    x_nb = x.multiply(r)
    return m.fit(x_nb, y), r

vec = TfidfVectorizer(ngram_range=(1,2), tokenizer=tokenize,
               min_df=3, max_df=0.9, strip_accents='unicode', use_idf=1,
               smooth_idf=1, sublinear_tf=1 )

# Model cross-validation w/ Stratified k-fold (within Toxic data set)

skf = StratifiedKFold(n_splits = 3, shuffle = True, random_state = 1)

for i, j in enumerate(label_cols):

    X = train[COMMENT]
    y = train[j]
    skf.get_n_splits(X,y)

    print('Model Stratified 3-Ford Cross-Validation for ', j,'\n')
    for train_index, test_index in skf.split(X, y):
        X_train0, X_test0 = X[train_index], X[test_index]
        y_train, y_test = y[train_index], y[test_index]

        X_train = vec.fit_transform(X_train0)
        X_test = vec.transform(X_test0)

        m,r = get_mdl(X_train, y_train)
        preds = m.predict_proba(X_test.multiply(r))[:,1]

        print(classification_report(y_test, np.round(preds)))

# Re-Model w/ all Toxic Train Data and Apply to Yelp Reviews

## Part 1. Fit to recommended file

n = train.shape[0]

x = vec.fit_transform(train[COMMENT])

test_x = vec.transform(regular_reviews['review'])

preds = np.zeros((len(regular_reviews), len(label_cols)))
for i, j in enumerate(label_cols):
    print('fit', j)
    m,r = get_mdl(x, train[j])
    preds[:,i] = m.predict_proba(test_x.multiply(r))[:,1]
    
regular_reviews = pd.concat([regular_reviews, pd.DataFrame(preds, columns = label_cols)], axis=1)
regular_reviews.to_csv('data/regular_NBSVM.csv', index=False)

## Part 2. Fit to not-recommended file

test_x = vec.transform(not_recommended_reviews['review'])

preds = np.zeros((len(not_recommended_reviews), len(label_cols)))
for i, j in enumerate(label_cols):
    print('fit', j)
    m,r = get_mdl(x, train[j])
    preds[:,i] = m.predict_proba(test_x.multiply(r))[:,1]

not_recommended_reviews = pd.concat([not_recommended_reviews, pd.DataFrame(preds, columns = label_cols)], axis=1)
not_recommended_reviews.to_csv('data/not_reviews_NBSVM.csv', index=False)

# 4. Readability

There are multiple readability score formulas as shown below.  They utilize count values of words, sentence, cyllables, and/or characters.  Each formular varies slightly, and gives out grade level who can read easily.  


 - Flesch-Kincaid
 - Coleman-Liau
 - Dale-Chall
 - SMOG
 - Automated Readability Index
 - Flesch Reading Ease (does not have min_age)

source <https://github.com/wimmuskee/readability-score>

### Readabiltiy Score Calculation

<https://en.wikipedia.org/wiki/Flesch%E2%80%93Kincaid_readability_tests>

https://en.wikipedia.org/wiki/Automated_readability_index

## Part 1. For recommended file

regular_reviews['readability_FK'] = regular_reviews.review.apply(lambda x:FleschKincaid(x).min_age)
regular_reviews['readability_AR'] = regular_reviews.review.apply(lambda x:ARI(x).min_age)
regular_reviews.to_csv('data/reg_reviews_Readability.csv', index=False)

## Part 2. For not-recommended file

not_recommended_reviews['readability_FK'] = not_recommended_reviews.review.apply(lambda x:FleschKincaid(x).min_age)
not_recommended_reviews['readability_AR'] = not_recommended_reviews.review.apply(lambda x:ARI(x).min_age)
not_recommended_reviews.to_csv('data/not_reviews_Readability.csv', index=False)

# 5. Spam classification

Spam Corpus Data : http://myleott.com/op-spam.html


## Objective :  Classifying Yelp review into spam index [1=spam, 0=not spam].


## Method: 

1. The raw training data set is processed to have a train and a valid data set.  
Code to process: <https://drive.google.com/drive/folders/1VkS0TkjoeQp-vLnXQsw25YeUEiIcbANe>

2. Logistic regression model will be tested with the corpus data itself to check the accuracy of model.

3. Merge corpus train and test set to model with all corpus data set (1600 recordings)

4. The trained model will be used to classify spam from the Yelp data set. 

train = pd.read_csv('data/train_SpamCorpus.csv')
valid = pd.read_csv('data/valid_SpamCorpus.csv')

# append corpus data sets
trainAll = train.append(valid, ignore_index=True)
len(trainAll)

# Modeling/Validation within Spam Corpus Data Set

##  <span style="color:blue">Crete a bag of words representation, as a term document matrix. 
##  <span style="color:blue">Use ngrams, as suggested in the NBSVM paper.

re_tok = re.compile(f'([{string.punctuation}“”¨«»®´·º½¾¿¡§£₤‘’])')
def tokenize(s): return re_tok.sub(r' \1 ', s).split()

### TF-IDF priors

n = train.shape[0]
vec = TfidfVectorizer(ngram_range=(1,2), tokenizer=tokenize,
               min_df=3, max_df=0.9, strip_accents='unicode', use_idf=1,
               smooth_idf=1, sublinear_tf=1 )

### Basic naive bayes feature equation:

def pr(x, y_i, y):
    p = x[y==y_i].sum(0)
    return (p+1) / ((y==y_i).sum()+1)

# Fit a model for one dependent at a time:
This function is designed for multiple dependent features/labels

def get_mdl(x, y):
    y = y.values
    r = np.log(pr(x,1,y) / pr(x,0,y))
    m = LogisticRegression(C=4, dual=True)
    x_nb = x.multiply(r)
    return m.fit(x_nb, y), r

trainAll.head()

# Model cross-validation w/ Stratified k-fold (within spam corpus data set)

skf = StratifiedKFold(n_splits = 3, shuffle = True, random_state = 1)

X = trainAll['review']
y = trainAll['class']
skf.get_n_splits(X,y)

print('Model Stratified 3-Ford Cross-Validation:')
for train_index, test_index in skf.split(X, y):
    X_train0, X_test0 = X[train_index], X[test_index]
    y_train, y_test = y[train_index], y[test_index]
    
    X_train = vec.fit_transform(X_train0)
    X_test = vec.transform(X_test0)
    
    m,r = get_mdl(X_train, y_train)
    preds = m.predict_proba(X_test.multiply(r))[:,1]

    print(classification_report(y_test, np.round(preds)))

# Re-Modeling w/ all 1600 spam corpus reviews to apply to Yelp reviews

n = trainAll.shape[0]
vec = TfidfVectorizer(ngram_range=(1,2), tokenizer=tokenize,
               min_df=3, max_df=0.9, strip_accents='unicode', use_idf=1,
               smooth_idf=1, sublinear_tf=1 )
x = vec.fit_transform(trainAll.review)

m,r = get_mdl(x,trainAll['class'])

# 4a. Prediction for recommended reveiws

test_x = vec.transform(regular_reviews['review'])
regular_reviews['spam'] = m.predict_proba(test_x.multiply(r))[:,1]
regular_reviews.to_csv('data/reg_reviews_full_NLP.csv', index=False)

# 4b. Predictions for not-recommended reviews

test_x = vec.transform(not_recommended_reviews['review'])
not_recommended_reviews['spam'] = m.predict_proba(test_x.multiply(r))[:,1]
not_recommended_reviews.to_csv('data/not_reviews_full_NLP.csv', index=False)
